Added payout types request as private as well
